refactor(@uform/next): next & antd use some type
