+++ packages/antd/package.json
@@ -29,6 +29,7 @@
   },
   "dependencies": {
     "@uform/react": "^0.3.6",
+    "@uform/types": "^0.3.6",
     "@uform/utils": "^0.3.6",
     "classnames": "^2.2.6",
     "moveto": "^1.7.4",
