--- packages/antd/src/types/components/button.d.ts
@@ -1,5 +1,0 @@
-import { BaseButtonProps } from 'antd/lib/button/button'
-
-export interface ISubmitProps extends Omit<BaseButtonProps, 'loading'> {
-  showLoading?: boolean
-}
