+++ packages/next/package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "@uform/react": "^0.3.6",
     "@uform/utils": "^0.3.6",
+    "@uform/types": "^0.3.6",
     "classnames": "^2.2.6",
     "moveto": "^1.7.4",
     "react-stikky": "^0.1.15",
